single jest config
